Finished plot titles.
Finished the filename parsing that gets the Piece info which is used for the titles. Added `update_info` to handle the info. Finished the titles themselves. Fixed `append_mean_and_dev` so that it updates the mean and std tags instead of duplicating them. Switched tube and plate plot colors. `Numbers` now saves a second .xlsx output file that is compatible with Office 365.