Shifting energy-z histogram z-limits.
The z-limits of the energy-z histograms now shift o match the shifting plates.
To do this, the 'plates' item was added to `info`, and `__get_e_lims` was changed to `__get_e_z_lims` and now also returns the z-limits.
Also added a couple TODOs regarding the powerpoint documentation of the results.