Titled and completed energy-xy plots. Increased transparency for the multi plots.
Added `plot_means` to the `EnergyVsXY` histograms. `Event` and `Run` now plot and save their energy-xy plots.
Also finished the energy-xy plot titles and made the plot aspect ratios 1:1.
The transparency of the lines in the energy-z multi plots now increases with the number of lines.
Fixed `go` in `pieces` to only save the overall (concatenated) DataFrame when `out_dir` is given.