Added beam offset in the simulation and analysis. Fixed swapped x and y axes.
Offset the beam in FCalPrimaryGeneratorAction.cc so it is in line with the "upper right" (+x and +y) tube.
Fixed the issue of swapped x and y axes in the energy-xy plots and the single tube energy deposit calculations. Now the histogram results are transposed immediately, and the `__tubes` method of `Numbers` is corrected to match the axes.
Run now parses the folder name to set the 'offset_beam' key in its `info`. `_get_titles` in `Piece` now adds 'Offset e-.' to the title when 'offset_beam' evaluates as `True`.
Changed the alpha (transparency) of the energy-z multi plots back to a constant. It's now 0.2.
Made plot titles wrap to 34 characters width (this doesn't function yet).