Finished important TODOs.
The version in this/the previous commit was used for the 2nd (fixed) version of the showers powerpoint as well as the offset beams powerpoint.
Removed finished TODO items, which include all of the TODOs for the showers powerpoint.
Fixed up the format of the documentation for `Piece`.
Removed the old tungsten plate series code in FCalDetectorConstruction.cc, which had been commented out.

Signed-off-by: Anson <anson@email.arizona.edu>