Rotatable tube calorimeters.
Added the option to rotate the tube calorimeters about the y-axis. The tube cals, and the holes in the PEEK box which house them, are rotated within the PEEK box.
Now, the 4 tube cals are placed in an assembly solid, and the PEEK box and 4-tube assembly are placed in a second "assembly box" before being added to the world.
`G4UnionSolid` is used to create the holes in the PEEK box, since `G4MultiUnion` leads to a "segmentation fault (core dumped)" error when it is subtracted from a solid and the resulting `G4SubtractionSolid` is placed in the world.

Signed-off-by: Anson <anson@email.arizona.edu>