Treated hits data for rotated tube calorimeters.
The tube cals are being rotated within the PEEK box about an axis in the y-direction through the center of the top right tube cal (sign of rotation goes by the RHR). For now, the analysis reverses the rotation of the hits within the PEEK box and cuts off any hits that end up outside of the PEEK box. This is done in `Events`. The data is kept as one `DataFrame`.

Updated `__main__.py` to handle command line arguments using the `argparse` package. The allowed arguments are the run directories, the output folder, and the energy of the runs.

Signed-off-by: Anson <anson@email.arizona.edu>