Fixed/updated materials and more geometry. Omitted data for all but one tube in analysis.
Fixed more geometry:
 Changed the three outer tube electrodes to copper.
 Corrected the material of the interior of the cryostat to vacuum and added a dead LAr gap between the frontmost plate and the inner cryostat wall.
 The plate-to-cryo distance has not been changed yet (although the current figure differs from the one calculated yesterday with Rob's help).
 Corrected overlapping tube cal holes in PEEK; radii of the holes are now 2.5 microns larger than the largest radius of the tubes (caused a few percent difference in simulation analysis results).
In FCalDetectorConstruction.cc, tabs were replaced by spaces and trailing spaces at the end of lines were removed, via copy and paste.
For 'offset' Events:
 All data within the PEEK box that lies outside of the singled out tube calorimeter is now cut out.
 Also, the e dep in the middle section of the top right tube cal, rather than the sum over the sections of the 4 tubes, is displayed on the energy-z plots. These two values should be the same, however, since the data for the other three tubes is cut out.
'offset_beam' and rotation of the tube data are now handled separately in `Event.__treat_hits`.
Moved `z_lims` to calc.py and widened them to account for the new box-PEEK gaps and the wider existing plate cal gaps.
Changed tube energy plot limits from a factor of 15 to a factor of 25 smaller than the plate energy limits. Changed overall energy limits to 400 and 650 MeV for the runs with 200 and 350 GeV incident electrons respectively.
Removed the unused `split_plot` function in `calc`, which was a duplicate of `__split_plot` in `calc.EnergyVsZ`.

Signed-off-by: Anson <anson@email.arizona.edu>