"Cleaned up" single tube version.
Cut the 4 tube cals down to a single centered tube cal. Removed code specific to the 4 tube setup.
Removed the extra x-direction width from the PEEK box since it is no longer needed to enclose the single tube at any angle. The PEEK dimensions now match the actual current values (40x40x43.5 mm).
Removed the offset in the incident beam and reorganized a bit of code in FCalPrimaryGeneractionAction.cc.
Replaced the `tubHoleRadius` parameter in the detector construction with `tubeHoleGap`. The radius of the hole in the PEEK box is larger than the tube outer radius by this amount.
Distance from the front tungsten plate to the cryostat (along the line through the center of the plate and perpendicular to it) remains at the original 24.27 mm and has not been changed to 24.836 mm.

Signed-off-by: Anson <anson@email.arizona.edu>