New plates convention with plot lims, wrapped titles, updated output file conventions.
Changed the convention that the analysis uses to recognize folder names for runs with X by Y plate setups to 'XbyY'. For example, '7by5' indicates 7 by 5 plates. The plate numbers can be any integers. The old convention of using '8-4', '7-5', and '6-6' in the folder names is still supported.
Updated the energy-z plot z-limits to match any integer number of plates. Also increased the `_plate_separation` used to shift these limits. It now reflects the geometry with 4.15 mm plate gaps (including the Kapton sheets which are not simulated) and 3.5 mm plates.
Fixed `Piece.__get_titles` so that it returns the titles with wrapped text.
Added 'z' to all energy-z plot file suffixes.
Updated `pieces.go`:
 Now makes a new subfolder for all of the output from each run.
 Now prints the full traceback of errors it encounters in the runs.
 Turned off output of the 'all-runs' files.

Signed-off-by: Anson <anson@email.arizona.edu>