Analysis of tube and beam shifts, cleaned up plot layouts with wrapped titles.
Added 'tube_shift' and 'beam_shift' options to `info`. The `parse_number` function was added to parse the numbers from filenames for these two options as well as for the 'tubes_angle'.
`Event.__treat_hits` now reverses shifts in the tube cal(s). `__treat_hits` was slightly reorganized so that `PEEK_positions` is now used for all the manipulation of the x, y and z data.
Updated `Piece.__get_titles`:
 Added plot title parts for the tube shift and beam shift.
 Set the wrapping length of the plot titles for z and xy plots separately and adjusted them to match the plot sizes. The xy plot wrapping needed careful adjustment.
 Fixed spacing of '$' symbols in the 'tubes_angle' title part, so the LaTeX part of the titles does not wrap (it doesn't display as LaTeX if there is a newline within the LaTeX).
 Reordered the title parts so the incident energy is after the plate numbers and other info.
 Removed documentation stating that the energy-z and energy-xy titles are the same.
Adjusted xy plot layouts in `EnergyVsXy._make_fig_and_axes` so that the layouts are static and consistent, and so the wrapped titles are positioned correctly above the plot. (This was done by manually setting the `rect` parameter in `Figure.tight_layout`.)

Signed-off-by: Anson <anson@email.arizona.edu>